Fix saving data in handleCreate function
We are now saving the form values in a combined state object and not
separately anymore. The "quick syntax" of "name" being name:"name" of
now needs to be replaced by a full object definition syntax (name:
formFields.name)

Furthermore streamline values access to as well use dot-notation further
up in that function.